Move storage service abstract class.
